core/services/runpod_handler.py
"""
RunPod serverless handler for model inference

This file would be deployed to RunPod to handle inference requests.
It provides the interface between RunPod and our models.
"""

import os
import sys
import json
import base64
import logging
from typing import Dict, Any
from io import BytesIO
import torch
from PIL import Image

# RunPod serverless handler
import runpod

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Model cache
_model_cache = {}

def load_model(model_type: str, model_name: str):
    """Load model into memory with caching"""
    cache_key = f"{model_type}:{model_name}"
    
    if cache_key in _model_cache:
        return _model_cache[cache_key]
    
    logger.info(f"Loading model: {model_type} - {model_name}")
    
    if model_type == "image":
        if "flux" in model_name.lower():
            from diffusers import FluxPipeline
            pipe = FluxPipeline.from_pretrained(
                f"black-forest-labs/{model_name}",
                torch_dtype=torch.float16
            ).to("cuda")
        else:
            from diffusers import StableDiffusionXLPipeline
            pipe = StableDiffusionXLPipeline.from_pretrained(
                model_name,
                torch_dtype=torch.float16
            ).to("cuda")
        _model_cache[cache_key] = pipe
        return pipe
        
    elif model_type == "3d":
        # Load HunYuan3D model
        # This would be the actual model loading code
        logger.info(f"Loading 3D model: {model_name}")
        # Placeholder for actual model loading
        return {"model": model_name, "loaded": True}
        
    else:
        raise ValueError(f"Unknown model type: {model_type}")

def generate_image(job):
    """Generate image using the specified model"""
    job_input = job["input"]
    
    # Extract parameters
    model_name = job_input.get("model_name", "FLUX.1-schnell")
    prompt = job_input["prompt"]
    negative_prompt = job_input.get("negative_prompt", "")
    width = job_input.get("width", 1024)
    height = job_input.get("height", 1024)
    steps = job_input.get("steps", 20)
    guidance_scale = job_input.get("guidance_scale", 7.5)
    seed = job_input.get("seed", -1)
    
    # Load model
    pipe = load_model("image", model_name)
    
    # Set seed if provided
    if seed > 0:
        generator = torch.Generator("cuda").manual_seed(seed)
    else:
        generator = None
    
    # Generate image
    logger.info(f"Generating image with prompt: {prompt[:50]}...")
    
    # Update progress
    runpod.serverless.progress_update(job, 10, "Loading model...")
    
    image = pipe(
        prompt=prompt,
        negative_prompt=negative_prompt,
        width=width,
        height=height,
        num_inference_steps=steps,
        guidance_scale=guidance_scale,
        generator=generator
    ).images[0]
    
    runpod.serverless.progress_update(job, 90, "Processing output...")
    
    # Convert to base64
    buffer = BytesIO()
    image.save(buffer, format="PNG")
    image_base64 = base64.b64encode(buffer.getvalue()).decode()
    
    return {
        "image": image_base64,
        "width": width,
        "height": height,
        "model": model_name,
        "seed": seed if seed > 0 else "random"
    }

def generate_3d(job):
    """Generate 3D model using HunYuan3D"""
    job_input = job["input"]
    
    # Extract parameters
    model_name = job_input.get("model_name", "hunyuan3d-2.1")
    image_base64 = job_input.get("image")
    prompt = job_input.get("prompt", "")
    quality_preset = job_input.get("quality_preset", "standard")
    
    # Update progress
    runpod.serverless.progress_update(job, 10, "Loading 3D model...")
    
    # Load model
    model = load_model("3d", model_name)
    
    # Decode image if provided
    if image_base64:
        image_data = base64.b64decode(image_base64)
        image = Image.open(BytesIO(image_data))
    else:
        # Generate image from prompt first
        runpod.serverless.progress_update(job, 20, "Generating reference image...")
        # Use image generation pipeline
        # ... (image generation code)
        image = None  # Placeholder
    
    runpod.serverless.progress_update(job, 50, "Converting to 3D...")
    
    # Run 3D generation
    # This would be the actual HunYuan3D inference code
    logger.info(f"Generating 3D model with quality: {quality_preset}")
    
    # Placeholder result
    model_data = b"GLB_FILE_DATA_HERE"  # This would be actual GLB data
    model_base64 = base64.b64encode(model_data).decode()
    
    runpod.serverless.progress_update(job, 90, "Finalizing model...")
    
    return {
        "model": model_base64,
        "format": "glb",
        "quality": quality_preset,
        "vertices": 10000,  # Placeholder
        "faces": 20000      # Placeholder
    }

def handler(job):
    """
    Main RunPod serverless handler
    
    Expected input format:
    {
        "input": {
            "model_type": "image" | "3d",
            "model_name": "model-identifier",
            ... (model specific parameters)
        }
    }
    """
    try:
        job_input = job["input"]
        model_type = job_input.get("model_type", "image")
        
        logger.info(f"Processing {model_type} generation request")
        
        if model_type == "image":
            result = generate_image(job)
        elif model_type == "3d":
            result = generate_3d(job)
        else:
            raise ValueError(f"Unknown model type: {model_type}")
        
        return result
        
    except Exception as e:
        logger.error(f"Error processing job: {str(e)}")
        return {"error": str(e)}

# RunPod serverless entrypoint
runpod.serverless.start({"handler": handler})