pyproject.toml
[project]
name = "hunyuan3d-app"
version = "0.1.0"
requires-python = ">=3.10,<3.12"
dependencies = [
    # Core - More flexible compatible versions
    "gradio>=5.0.0", # Use newer version to avoid urllib3 conflicts
    # PyTorch - Comment out and install manually with correct CUDA version
    # "torch>=2.4.0,<2.5.0",  # Pin torch version for faster install
    # "torchvision>=0.19.0,<0.20.0",
    # "torchaudio>=2.4.0,<2.5.0",
    # ML/AI - Compatible versions
    "transformers>=4.46.1", # Fixed version for Python 3.8+ compatibility
    "diffusers>=0.32.0", # Supports GGUF model loading
    "accelerate>=0.32.0",
    "huggingface-hub>=0.19.0",
    "hf_transfer>=0.1.0", # For faster HuggingFace downloads
    "safetensors>=0.4.0",
    "sentencepiece",
    "gguf>=0.10.0", # For GGUF model support
    "ninja", # Faster CUDA kernel compilation
    "nvitop>=1.3.0", # GPU monitoring
    "timm>=1.0.0", # Vision transformer models for Hunyuan3D
    # 3D Processing
    "trimesh>=4.0.0",
    "pymeshlab",
    "open3d>=0.17.0",
    "scikit-image>=0.21.0",
    "point-cloud-utils>=0.29.0",
    "pyglet<2.0.0", # For 3D rendering and preview generation (trimesh requires <2.0)
    "xatlas>=0.0.9", # For UV unwrapping in HunYuan3D texture generation
    # Image Processing
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "rembg>=2.0.50",
    "imageio>=2.31.0",
    "imageio-ffmpeg>=0.4.9",
    # Utils
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "einops>=0.7.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "tensorboard>=2.14.0",
    "psutil>=5.9.0", # Added for system monitoring
    "pywin32>=228; platform_system == 'Windows'", # For Windows-specific functionality
    "pynvml>=11.5.0", # For NVIDIA GPU detection
    "gputil>=1.4.0", # For GPU monitoring
    "controlnet-aux>=0.0.7", # For ControlNet preprocessing (depth, canny, pose, etc.)
    # Security & API
    "keyring>=24.0.0", # Secure credential storage
    "cryptography>=41.0.0", # Encryption for fallback storage
    "aiohttp>=3.9.0", # Async HTTP for Civitai downloads
    # New advanced features
    "websockets>=12.0", # WebSocket server for real-time progress
    "python-socketio>=5.10.0", # Socket.IO support
    "scikit-learn>=1.3.0", # For NLP and clustering in LoRA suggestions
    "clip-interrogator>=0.6.0", # For image analysis in character consistency
    "onnxruntime>=1.16.0", # For face swapping and InsightFace models
    "aria2p>=0.11.0", # Fast downloads for Civitai
    # FaceFusion 3.2.0 dependencies
    "onnxruntime-gpu>=1.16.0; platform_system != 'Darwin'", # GPU acceleration for non-Mac
    "opencv-contrib-python>=4.8.0", # Extended OpenCV features for FaceFusion
    "insightface>=0.7.3", # Latest InsightFace for better face detection
    "gfpgan>=1.3.8", # Face restoration models
    "realesrgan>=0.3.0", # Super resolution for face enhancement
    # Video generation placeholders (actual packages would be added when available)
    # "ltx-video>=0.1.0",  # LTX-Video model
    # "wan-video>=2.1.0",  # Wan 2.1 model
    # "skyreels>=0.1.0",  # SkyReels model
    # Memory optimizations (optional for quick testing)
    # "xformers>=0.0.28; platform_system == 'Windows'",  # Windows-compatible xformers
    # "xformers>=0.0.22; platform_system == 'Linux'",    # Linux xformers
    "nicegui>=1.4.0",
    "ollama>=0.1.0",
    "pyqtgraph>=0.13.3",
    "qasync>=0.24.0",
    "pywebview>=5.4",
    "mediapipe>=0.10.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

# FaceFusion advanced features
facefusion = [
    "torch>=2.1.0",  # Required for FaceFusion
    "torchvision>=0.16.0",
    "onnx>=1.15.0",
    "protobuf>=4.21.0",
    "tqdm>=4.65.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
]

# Windows-specific optimizations
windows = [
    # Additional Windows-specific packages can go here
]


# Linux-specific optimizations
linux = [
    "bitsandbytes>=0.41.0; sys_platform == 'linux'",
    "flash-attn>=2.5.0; sys_platform == 'linux'",
    "triton>=2.1.0; sys_platform == 'linux'",
]

# Additional CUDA-specific optimizations
cuda118 = [
    "flash-attn>=2.0.0; sys_platform == 'linux'",  # Flash attention for Linux only
]

cuda121 = [
    "flash-attn>=2.5.0; sys_platform == 'linux'",  # Flash attention for Linux only
    "triton>=2.1.0; sys_platform == 'linux'",  # For torch.compile optimization
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hunyuan3d_app"]

[tool.uv]
# Use unsafe-best-match to resolve conflicts across indexes
index-strategy = "unsafe-best-match"


# Override dependencies that might have conflicts
override-dependencies = [
    "numba>=0.57.0",
    "llvmlite>=0.40.0",
    "rembg>=2.0.57",
    "urllib3>=2.0.0,<3.0.0",  # Force compatible urllib3 version
    "requests>=2.31.0",       # Use newer requests version
]

# Platform-specific constraints to prevent cross-platform issues
constraint-dependencies = [
    "triton>=2.1.0; sys_platform == 'linux'",  # Ensure triton only on Linux
    "bitsandbytes>=0.41.0; sys_platform == 'linux'",  # Linux only
    "flash-attn>=2.0.0; sys_platform == 'linux'",  # Linux only
]

# Extra index URLs for PyTorch and xformers
# CUDA 12.4+ compatible (includes CUDA 12.6)
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = 10

# Xformers Windows binaries
[[tool.uv.index]]
name = "xformers"
url = "https://download.pytorch.org/whl/cu124"
priority = 10
